Rename `:std:visibility-lint` to ':std:visibility-check'
