+++ std/visibility-check/src/main/java/br/com/orcinus/orca/std/visibility/check/detection/UDeclaration.extensions.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright © 2024 Orcinus
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+ * the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program. If
+ * not, see https://www.gnu.org/licenses.
+ */
+
+package br.com.orcinus.orca.std.visibility.check.detection
+
+import br.com.orcinus.orca.std.visibility.PackageProtected
+import com.android.tools.lint.detector.api.JavaContext
+import com.intellij.psi.PsiAnnotation
+import org.jetbrains.uast.UAnnotation
+import org.jetbrains.uast.UDeclaration
+import org.jetbrains.uast.toUElementOfType
+
+/**
+ * Obtains the [UAnnotation] of [PackageProtected] if this [UDeclaration] is of a structure that has
+ * been annotated with it or with a [UAnnotation] that extends [PackageProtected], denoting that
+ * references from a package that isn't the one in which this [UDeclaration] is should be reported.
+ *
+ * @param context [JavaContext] for finding [UAnnotation]s that might be
+ *   [PackageProtected]-annotated, allowing for improper references to be propagated with
+ *   [UAnnotation]s other than [PackageProtected] itself.
+ */
+internal fun UDeclaration.findPackageProtectedAnnotation(context: JavaContext): UAnnotation? {
+  return with(PackageProtected::class.java.name) {
+      getAnnotation(this)
+        ?: annotations
+          .mapNotNull(PsiAnnotation::getQualifiedName)
+          .mapNotNull(context.evaluator::findClass)
+          .firstNotNullOfOrNull { it.getAnnotation(this) }
+    }
+    ?.toUElementOfType()
+}
